moved ModeDropdown out of NavBar
