+++ src/components/Navigation/ModeDropdown.scss
@@ -1,0 +1,4 @@
+.gadgetContainer {
+  display: inline-block;
+  padding: 3px 5px;
+}
